Added support for Scala 2.13

+++ squbs-pattern/src/main/scala/org/squbs/pattern/timeoutpolicy/TimeoutPolicy.scala
@@ -71,9 +71,8 @@ abstract class TimeoutPolicy(name: Option[String], initial: FiniteDuration, star
         logger.warn("call end without call waitTime first, ignore this transaction")
       } else {
         val isTimeout = timeTaken > waitTime.toNanos
-        TimeoutPolicy.this.update(timeTaken, isTimeout)
+        TimeoutPolicy.this.update(timeTaken.toDouble, isTimeout)
       }
-
     }
   }
